pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "codescanner"
version = "1.0.0"
description = "Multi-tool security vulnerability scanner with AI analysis"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "CodeScanner Team"}
]
keywords = ["security", "vulnerability", "scanner", "static-analysis", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: X11 Applications",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "google-genai>=1.0.0",
    "customtkinter>=5.2.0",
    "pathspec>=0.12.0",
    "aiofiles>=24.0.0",
]

[project.optional-dependencies]
all = [
    "bandit>=1.7.0",
    "semgrep>=1.50.0",
    "safety>=3.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
    "ruff>=0.4.0",
]

[project.scripts]
codescanner = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "S", "SIM", "RUF"]
ignore = ["E501", "S101", "S603", "S607"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
